Move getUnparsedFlagValue to cli/util/flags.go and export it.
