+++ pkg/client/cli/intercept/state.go
@@ -328,35 +328,6 @@ func (s *state) CreateRequest(ctx context.Context) (*connector.CreateInterceptRe
 	return ir, nil
 }
 
-// getUnparsedFlagValue returns the value of a flag that has been provided after a "--" on the command
-// line, and hence hasn't been parsed as a normal flag. Typical use case is:
-//
-//	telepresence intercept --docker-run ... -- --name <name>
-func getUnparsedFlagValue(args []string, flag string) (string, error) {
-	feq := flag + "="
-	for i, arg := range args {
-		var v string
-		switch {
-		case arg == flag:
-			i++
-			if i < len(args) {
-				if v = args[i]; strings.HasPrefix(v, "-") {
-					v = ""
-				}
-			}
-		case strings.HasPrefix(arg, feq):
-			v = arg[len(feq):]
-		default:
-			continue
-		}
-		if v == "" {
-			return "", fmt.Errorf("flag %q requires a value", flag)
-		}
-		return v, nil
-	}
-	return "", nil
-}
-
 func (s *state) startInDocker(ctx context.Context, envFile string, args []string) (*dexec.Cmd, error) {
 	ourArgs := []string{
 		"run",
@@ -364,7 +335,7 @@ func (s *state) startInDocker(ctx context.Context, envFile string, args []string
 		"--dns-search", "tel2-search",
 	}
 
-	name, err := getUnparsedFlagValue(args, "--name")
+	name, err := util.GetUnparsedFlagValue(args, "--name")
 	if err != nil {
 		return nil, err
 	}
